+++ modules/pytket-braket/docs/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 * Allow user to specify different local backends.
 * Fix two-qubit unitary conversions.
+* Fix dimension mismatch for measurements for Braket QPU results
 
 0.17.0 (March 2022)
 -------------------